requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-webpack-loader==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
